Introduce a delta_apply security label management.
With this commit, we add the 'spock.delta_apply' routine, which adds or removes
a security label on a specific table column.
It is mostly useless for now. It demonstrates how to consistently manage
security labels and survive ALTER/DROP DDL commands, including column
type change.
Trivial regression tests are added to fix the appropriate behaviour.