Use security labels to implement the delta_apply feature.
The general idea is to utilise the REPLICA_DENTITY_FULL status to force
the backend to write the WAL record columns we need to calculate a delta tuple.
Auto DDL replication transfers security labels (and replica identity changes)
across all nodes involved. On the subscriber side, the SpockRelation cache
contains a specific Oid list for each column, or an InvalidOid if the column
doesn't need the delta_apply feature.

This commit doesn't contain any tests, nor does it remove the previous 'core'
approach; that should be done in the following commits.

Words on upgrade: there is no easy way to upgrade from a patched version to
the extensible one. So, at a glance, it seems we could just recommend that
people remove the delta apply settings before the upgrade and add them after
the upgrade using the new interface.

NOTE:
There is no good reason to deliver the whole tuple by the LR protocol to
the subscriber. We can send only IDENTITY columns and columns needed for
the delta_apply. It may be done by employing SpockRelayion cache inside
the walsender.

TODO:
To stay safe, we should limit 'delta apply' types with increment, decrement
and initialisation operators only. Also, we should decline nullable columns.