Separate the long-lasting Z0DAN sync TAP test.
Designed to be highly competitive, this test consumes more CPU, disk, and
memory resources than usual. To let people pass basic checks on their ongoing
commits faster, turn off this test in the schedule and separate it as a GitHub
action.

Also, it appears that we copied Spock into the base image. Hence, the 'step-1'
Docker image might use oldish versions of the patches.