Improve the 011_zodan_sync_third TAP test.
There are some corner cases that, at least for now, need to involve both
the publisher and the subscriber when we try to determine whether all the data
is replicated (synchronisation has finished). So, rewrite the sync code
to sync_event/wait calls.

It seems that if we improve wait functions a little, it may serve to
multiple purposes. For example, we may add sync_event to the wait function,
replace integer timeouts with the Interval type, and introduce 'verbose' mode.
And yes, rewrite it as a function.