Remove the SPI API support from the Spock protocol.
It appears to be a legacy feature that has never been utilised.
It is unsupported, and it seems it was never supported, even in the project's
predecessor, 'pglogical.'

By making this change, we can simplify the codebase, emphasise the key
features, and streamline development. If necessary, this change can be quickly
reversed.