Remove support for the JSON message format in the Spock protocol.
It is a legacy that has never been supported and tested. Looking around, we can
see that it is not supported in pglogical either and has been removed
from pgactive.

Hence, for simplicity and codebase stability, this feature should be removed
from the protocol. Anyway, it may be reverted fast if needed.