Further refinement of the RM feature.
Previously, the adjust_progress_info routine updated the progress before
the COPY transaction was committed. This approach was acceptable because
if the transaction aborted, we could simply roll back all the progress changes.
However, now that we are using an HTAB, we report progress after the commit
(as seen in the handle_commit routine). Therefore, the spock_sync module
should follow the same procedure.

Additionally, this commit introduces the local_lsn field in the progress output
and implements an interface check in the regression tests. This new field
indicates the exact LSN position at which this progress record is stored
locally. We will use it later to simplify and optimise the search for
the correct remote LSN after the initial subscription synchronisation in
configurations with 3+ nodes.

NOTE:
Minor change was made in the sync_table test to let it be less dependent on
the usage history of the spock status table.