Change the progress arrangement logic when syncing a subscription.
The COPY procedure may take a long time. So, when loading the progress state
after the COPY on an actively processing transactions Spock 3+ configuration,
there's a risk that more transactions from ‘other’ nodes have arrived and
been applied on the ‘donor’ node.

So, if someone wants to continue LR from the ‘other’ nodes without conflicting
with data copied from the ‘donor’ node, we need to calculate the precise state
of replication progress consistent with the (snapshot, LSN) returned by
the CREATE_REPLICATION_SLOT syncing operation.

Taking the progress state right before creating a slot narrows the window
during which ‘other’ transactions may arrive and be committed by apply workers.
For sure, it is not an ultimate solution, but a step in the right direction.
The following steps are:

1. Write a test that shows the problem may evolve if something extends the gap
in the subscription sync procedure (probably using injection points?).
2. Write a function that scans the gap and updates the progress state.

NOTE:
This commit doesn’t change the table sync code. It is not clear whether we need
to arrange the entire LR progress on an N3 node to the current state of
a ‘donor’ node N1, since someone decided to resync a single table. So, let it
behave as usual.