Check transactional state in spock_node.
There are some evidence exists that we have been cavalier with database
lookups. It seems sometimes Spock does it outside a transaction. And the reason
why we didn't see any issues before is that relation we touch is frequently
in the syscache. It is unstable situation and we must fix it.
It is too expensive to run tests with the '--discard-caches' enabled. So, to
make thing simpler, add assertions to check we touch the database inside
a transaction.