Check the IF EXISTS clause in add_ddl_to_repset.
In case of IF EXISTS / NOT EXISTS clause Postgres core utility call produces
an INFO message and lets utility hooks do their job as usual. In the Spock
case, we need to do the same and carefully process the clause so it doesn't
produce an ERROR that breaks the convention.