During subscription syncing, take progress info more consistently.
In Spock 5, we used to employ a progress table and a transactional snapshot
to ensure that the progress corresponds to the replication slot
(snapshot, LSN), and we could request this state at any time until
the transaction ends.

With an HTAB, it's no longer true, and we need to invent extra hacks. With this
commit, we take the progress state right before LR slot creation, which is as
close as possible to the real state.

It is still not entirely consistent in the case of Z0DAN, but while we use it
just for information, it is ok.

XXX:
The code related to table syncing stays as is. Right now, we aren't sure if
the progress state needs to be corrected if we re-sync a single table from
the database. So, just keep the code untouched.